# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: The Champions Club API
  version: 1.0.0
  description: "GENERATED from tools/api/endpoints.mjs — DO NOT EDIT BY HAND. Human-readable version: API_CONTRACT.md. Schemas are derived from shared/types/*.ts (TypeScript is authoritative; inline/complex shapes may appear as free-form)."
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      properties:
        success:
          description: "TS: false"
        error:
          type: object
          properties:
            code:
              description: "TS: ErrorCode"
            message:
              type: string
            details:
              type: object
              additionalProperties: {}
          required:
            - code
            - message
      required:
        - success
        - error
...
    Court:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        sport_type:
          type: string
          enum:
            - TENNIS
            - CRICKET
            - PADEL
            - BADMINTON
        description:
          type: string
          nullable: true
        surface:
          type: string
          nullable: true
        walk_in_rate_per_hour:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "1250.00"
        image_url:
          type: string
          nullable: true
        sort_order:
          type: number
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - sport_type
        - description
        - surface
        - walk_in_rate_per_hour
        - image_url
        - sort_order
        - is_active
        - created_at
        - updated_at
...
    MembershipView:
      allOf:
        - $ref: "#/components/schemas/Membership"
        - type: object
          properties:
            status:
              type: string
              enum:
                - UPCOMING
                - ACTIVE
                - EXPIRED
                - CANCELLED
            plan:
              $ref: "#/components/schemas/MembershipPlan"
            member_name:
              type: string
            member_code:
              type: string
          required:
            - status
            - plan
            - member_name
            - member_code
...
    MembershipPurchaseResult:
      type: object
      properties:
        membership:
          $ref: "#/components/schemas/MembershipView"
        payment:
          $ref: "#/components/schemas/Payment"
          nullable: true
      required:
        - membership
        - payment
...
    DateRange:
      type: object
      properties:
        from:
          type: string
          format: date
        to:
          type: string
          format: date
      required:
        - from
        - to
...
    RevenueReport:
      type: object
      properties:
        range:
          $ref: "#/components/schemas/DateRange"
        group_by:
          type: string
          enum:
            - DAY
            - CATEGORY
            - METHOD
        rows:
          type: array
          items:
            description: "TS: AmountByKey & { tax_amount: Money }"
        total:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "1250.00"
        total_tax:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "1250.00"
      required:
        - range
        - group_by
        - rows
        - total
        - total_tax
    CourtUtilizationReport:
      type: object
      properties:
        range:
          $ref: "#/components/schemas/DateRange"
        rows:
          type: array
          items:
            type: object
            properties:
              court_id:
                type: string
                format: uuid
              court_name:
                type: string
              booked_hours:
                type: number
              available_hours:
                type: number
              utilization_percent:
                type: string
                pattern: ^\d+\.\d{2}$
                example: "18.00"
              revenue:
                type: string
                pattern: ^\d+\.\d{2}$
                example: "1250.00"
              cancellations_count:
                type: number
            required:
              - court_id
              - court_name
              - booked_hours
              - available_hours
              - utilization_percent
              - revenue
              - cancellations_count
      required:
        - range
        - rows
    MembershipReport:
      type: object
      properties:
        range:
          $ref: "#/components/schemas/DateRange"
        active_by_type:
          type: array
          items:
            $ref: "#/components/schemas/AmountByKey"
        new_members:
          type: number
        renewals:
          type: number
        membership_revenue:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "1250.00"
        expiring:
          type: array
          items:
            $ref: "#/components/schemas/MemberSummary"
      required:
        - range
        - active_by_type
        - new_members
        - renewals
        - membership_revenue
        - expiring
...
paths:
...
  /members/{id}/memberships:
    get:
      operationId: members.memberships
      tags:
        - Members
      summary: All membership terms of a member (plan changes, renewals, expiries).
      description: "Requirements: FR-MEM-004, FR-MEM-005"
      x-roles:
        - MEMBER:own
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 2
      x-requirements:
        - FR-MEM-004
        - FR-MEM-005
      x-tables:
        - memberships
        - membership_plans
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MembershipView"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: MEMBER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /memberships:
    post:
      operationId: memberships.purchase
      tags:
        - Membership plans & memberships
      summary: Buy or renew a membership. Records the payment.
      description: |-
        No current term: new membership starting today (derived status ACTIVE). Current term ends within `membership_expiry_warning_days`: new term starts the day after the current end_date (derived status UPCOMING until then). Otherwise MEMBERSHIP_ALREADY_ACTIVE (the DB rejects overlapping live terms: memberships_no_overlap). end_date = start_date + duration_months - 1 day. A membership is only member + plan + start/end date + price_paid: its status is derived from the dates (view membership_terms), never stored.

        Requirements: FR-MEM-007, FR-MEM-012, FR-FIN-001
      x-roles:
        - MEMBER:own
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 2
      x-requirements:
        - FR-MEM-007
        - FR-MEM-012
        - FR-FIN-001
      x-tables:
        - memberships
        - payments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                member_id:
                  type: string
                  format: uuid
                  description: "Staff only; MEMBER callers: implied = self"
                membership_plan_id:
                  type: string
                  format: uuid
                payment_method:
                  type: string
                  enum:
                    - CASH
                    - CARD
                    - UPI
                    - ONLINE
                  description: MEMBER callers must use ONLINE
                gateway_reference:
                  type: string
              required:
                - membership_plan_id
                - payment_method
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/MembershipPurchaseResult"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "402":
          description: PAYMENT_FAILED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: MEMBER_NOT_FOUND | MEMBERSHIP_PLAN_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: MEMBERSHIP_ALREADY_ACTIVE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          description: JUNIOR_AGE_INVALID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /courts:
    get:
      operationId: courts.list
      tags:
        - Courts & availability
      summary: Court catalogue.
      description: "Requirements: FR-COURT-001, FR-PUB-003"
      x-roles:
        - PUBLIC
      x-owner: Dev 2
      x-requirements:
        - FR-COURT-001
        - FR-PUB-003
      x-tables:
        - courts
      security: []
      parameters:
        - name: sport_type
          in: query
          required: false
          schema:
            type: string
            enum:
              - TENNIS
              - CRICKET
              - PADEL
              - BADMINTON
        - name: include_inactive
          in: query
          required: false
          description: staff only
          schema:
            type: boolean
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Court"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: courts.create
      tags:
        - Courts & availability
      summary: Add a court.
      description: "Requirements: FR-COURT-015"
      x-roles:
        - OWNER_ADMIN
      x-owner: Dev 2
      x-requirements:
        - FR-COURT-015
      x-tables:
        - courts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: unique
                sport_type:
                  type: string
                  enum:
                    - TENNIS
                    - CRICKET
                    - PADEL
                    - BADMINTON
                description:
                  type: string
                surface:
                  type: string
                walk_in_rate_per_hour:
                  type: string
                  pattern: ^\d+\.\d{2}$
                image_url:
                  type: string
                  format: uri
                sort_order:
                  type: integer
              required:
                - name
                - sport_type
                - walk_in_rate_per_hour
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/Court"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /reports/dashboard:
    get:
      operationId: reports.dashboard
      tags:
        - Reports
      summary: Owner dashboard KPIs for today / this week / this month across every module.
      description: |-
        Revenue = SUM(payments.amount - refunded_amount) by paid_at in the range (IST), grouped by the derived revenue_category (view payment_ledger).

        Requirements: FR-REP-001, FR-FIN-009, FR-MEM-014
      x-roles:
        - OWNER_ADMIN
      x-owner: Dev 4
      x-requirements:
        - FR-REP-001
        - FR-FIN-009
        - FR-MEM-014
      x-tables:
        - payments
        - memberships
        - court_bookings
        - shop_orders
        - bar_orders
        - enquiries
        - invoices
        - payroll_payments
        - leave_requests
        - products
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          required: true
          description: TODAY = IST today; WEEK = Monday..today; MONTH = 1st..today
          schema:
            type: string
            enum:
              - TODAY
              - WEEK
              - MONTH
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/OwnerDashboard"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /reports/revenue:
    get:
      operationId: reports.revenue
      tags:
        - Reports
      summary: Revenue by day, category (court/membership/shop/bar/business) or payment method.
      description: "Requirements: FR-REP-002, FR-FIN-003, FR-FIN-002"
      x-roles:
        - OWNER_ADMIN
      x-owner: Dev 4
      x-requirements:
        - FR-REP-002
        - FR-FIN-003
        - FR-FIN-002
      x-tables:
        - payments
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          description: Inclusive IST business date.
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: Inclusive IST business date; to >= from; max 366 days.
          schema:
            type: string
            format: date
        - name: group_by
          in: query
          required: false
          description: default CATEGORY
          schema:
            type: string
            enum:
              - DAY
              - CATEGORY
              - METHOD
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/RevenueReport"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /reports/courts:
    get:
      operationId: reports.courts
      tags:
        - Reports
      summary: Court utilisation, cancellations and revenue per court.
      description: |-
        utilization_percent = booked hours / available hours (open hours x days, minus MAINTENANCE).

        Requirements: FR-REP-003
      x-roles:
        - OWNER_ADMIN
      x-owner: Dev 4
      x-requirements:
        - FR-REP-003
      x-tables:
        - court_bookings
        - courts
        - payments
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          description: Inclusive IST business date.
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: Inclusive IST business date; to >= from; max 366 days.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/CourtUtilizationReport"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /reports/memberships:
    get:
      operationId: reports.memberships
      tags:
        - Reports
      summary: Active members by type, new registrations, renewals, expiring, membership revenue.
      description: "Requirements: FR-REP-004, FR-MEM-014"
      x-roles:
        - OWNER_ADMIN
      x-owner: Dev 4
      x-requirements:
        - FR-REP-004
        - FR-MEM-014
      x-tables:
        - memberships
        - members
        - payments
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          description: Inclusive IST business date.
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: Inclusive IST business date; to >= from; max 366 days.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/MembershipReport"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /reports/shop:
    get:
      operationId: reports.shop
      tags:
        - Reports
      summary: Shop sales, fulfillment split, top products, low stock.
      description: "Requirements: FR-REP-005"
      x-roles:
        - OWNER_ADMIN
      x-owner: Dev 4
      x-requirements:
        - FR-REP-005
      x-tables:
        - shop_orders
        - shop_order_items
        - products
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          description: Inclusive IST business date.
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: Inclusive IST business date; to >= from; max 366 days.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/ShopReport"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /reports/bar:
    get:
      operationId: reports.bar
      tags:
        - Reports
      summary: Cafe revenue per day (what the cafe earned), by method.
      description: "Requirements: FR-REP-006, FR-BAR-009"
      x-roles:
        - OWNER_ADMIN
      x-owner: Dev 4
      x-requirements:
        - FR-REP-006
        - FR-BAR-009
      x-tables:
        - payments
        - bar_orders
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          description: Inclusive IST business date.
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: Inclusive IST business date; to >= from; max 366 days.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BarDailySummary"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /reports/finance:
    get:
      operationId: reports.finance
      tags:
        - Reports
      summary: Money in, refunds, outstanding invoices, overdue, payroll paid/pending, tax collected.
      description: "Requirements: FR-REP-009, FR-FIN-006, FR-FIN-008"
      x-roles:
        - OWNER_ADMIN
      x-owner: Dev 4
      x-requirements:
        - FR-REP-009
        - FR-FIN-006
        - FR-FIN-008
      x-tables:
        - payments
        - invoices
        - payroll_payments
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          description: Inclusive IST business date.
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: Inclusive IST business date; to >= from; max 366 days.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/FinanceReport"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /reports/tax:
    get:
      operationId: reports.tax
      tags:
        - Reports
      summary: Tax collected (GST) by revenue category for filing.
      description: |-
        Sum of payments.tax_amount net of refunds (pro-rata), grouped by the derived revenue_category.

        Requirements: FR-REP-007, FR-FIN-007
      x-roles:
        - OWNER_ADMIN
      x-owner: Dev 4
      x-requirements:
        - FR-REP-007
        - FR-FIN-007
      x-tables:
        - payments
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          description: Inclusive IST business date.
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: Inclusive IST business date; to >= from; max 366 days.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/TaxReport"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /reports/export:
    get:
      operationId: reports.export
      tags:
        - Reports
      summary: Download a report as CSV to share with an accountant / partners. Response is text/csv, NOT the JSON envelope.
      description: "Requirements: FR-REP-008"
      x-roles:
        - OWNER_ADMIN
      x-owner: Dev 4
      x-requirements:
        - FR-REP-008
      x-tables:
        - payments
        - court_bookings
        - members
        - shop_orders
        - bar_orders
      security:
        - bearerAuth: []
      parameters:
        - name: report
          in: query
          required: true
          schema:
            type: string
            enum:
              - REVENUE
              - MEMBERS
              - BOOKINGS
              - SHOP_SALES
              - BAR_SALES
              - TAX
        - name: from
          in: query
          required: true
          description: Inclusive IST business date.
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: Inclusive IST business date; to >= from; max 366 days.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: CSV download
          content:
            text/csv:
              schema:
                type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...